backend: put makeKmzFile to edit.js
